@@ -6,10 +6,13 @@ import re
 import time
 import warnings
 from abc import ABC, abstractmethod
-from datetime import date, datetime, timedelta
+from collections.abc import Iterable
+from datetime import datetime, timedelta, timezone
+from enum import Enum
 from pathlib import Path
-from typing import IO, Callable, Dict, Iterable, List, Optional, Union
+from typing import IO, Callable, Optional, Union
 
+import cloudscraper
 import numpy as np
 import pandas as pd
 import requests
@@ -17,9 +20,178 @@ import selenium
 import undetected_chromedriver as uc
 from dateutil.relativedelta import relativedelta
 from packaging import version
-from selenium.common.exceptions import WebDriverException
+from selenium.common.exceptions import JavascriptException, WebDriverException
 
 from ._config import DATA_DIR, LEAGUE_DICT, MAXAGE, logger
+
+
+class SeasonCode(Enum):
+    """How to interpret season codes.
+
+    Attributes
+    ----------
+    SINGLE_YEAR: The season code is a single year, e.g. '2021'.
+    MULTI_YEAR: The season code is a range of years, e.g. '2122'.
+    """
+
+    SINGLE_YEAR = "single-year"
+    MULTI_YEAR = "multi-year"
+
+    @staticmethod
+    def from_league(league: str) -> "SeasonCode":
+        """Return the default season code for a league.
+
+        Parameters
+        ----------
+        league : str
+            The league to consider.
+
+        Raises
+        ------
+        ValueError
+            If the league is not recognized.
+
+        Returns
+        -------
+        SeasonCode
+            The season code format to use.
+        """
+        if league not in LEAGUE_DICT:
+            raise ValueError(f"Invalid league '{league}'")
+        select_league_dict = LEAGUE_DICT[league]
+        if "season_code" in select_league_dict:
+            return SeasonCode(select_league_dict["season_code"])
+        start_month = datetime.strptime(  # noqa: DTZ007
+            select_league_dict.get("season_start", "Aug"),
+            "%b",
+        ).month
+        end_month = datetime.strptime(  # noqa: DTZ007
+            select_league_dict.get("season_end", "May"),
+            "%b",
+        ).month
+        return SeasonCode.MULTI_YEAR if (end_month - start_month) < 0 else SeasonCode.SINGLE_YEAR
+
+    @staticmethod
+    def from_leagues(leagues: list[str]) -> "SeasonCode":
+        """Determine the season code to use for a set of leagues.
+
+        If the given leagues have different default season codes,
+        the multi-year format is usded.
+
+        Parameters
+        ----------
+        leagues : list of str
+            The leagues to consider.
+
+        Returns
+        -------
+        SeasonCode
+            The season code format to use.
+        """
+        season_codes = {SeasonCode.from_league(league) for league in leagues}
+        if len(season_codes) == 1:
+            return season_codes.pop()
+        warnings.warn(
+            "The leagues have different default season codes. Using multi-year season codes.",
+            stacklevel=2,
+        )
+        return SeasonCode.MULTI_YEAR
+
+    def parse(self, season: Union[str, int]) -> str:  # noqa: C901
+        """Convert a string or int to a standard season format."""
+        season = str(season)
+        patterns = [
+            (
+                re.compile(r"^[0-9]{4}$"),  # 1994 | 9495
+                lambda s: process_four_digit_year(s),
+            ),
+            (
+                re.compile(r"^[0-9]{2}$"),  # 94
+                lambda s: process_two_digit_year(s),
+            ),
+            (
+                re.compile(r"^[0-9]{4}-[0-9]{4}$"),  # 1994-1995
+                lambda s: process_full_year_range(s),
+            ),
+            (
+                re.compile(r"^[0-9]{4}/[0-9]{4}$"),  # 1994/1995
+                lambda s: process_full_year_range(s.replace("/", "-")),
+            ),
+            (
+                re.compile(r"^[0-9]{4}-[0-9]{2}$"),  # 1994-95
+                lambda s: process_partial_year_range(s),
+            ),
+            (
+                re.compile(r"^[0-9]{2}-[0-9]{2}$"),  # 94-95
+                lambda s: process_short_year_range(s),
+            ),
+            (
+                re.compile(r"^[0-9]{2}/[0-9]{2}$"),  # 94/95
+                lambda s: process_short_year_range(s.replace("/", "-")),
+            ),
+        ]
+
+        current_year = datetime.now(tz=timezone.utc).year
+
+        def process_four_digit_year(season: str) -> str:
+            """Process a 4-digit string like '1994' or '9495'."""
+            if self == SeasonCode.MULTI_YEAR:
+                if int(season[2:]) == int(season[:2]) + 1:
+                    if season == "2021":
+                        msg = (
+                            f'Season id "{season}" is ambiguous: '
+                            f'interpreting as "{season[:2]}-{season[-2:]}"'
+                        )
+                        warnings.warn(msg, stacklevel=1)
+                    return season
+                if season[2:] == "99":
+                    return "9900"
+                return season[-2:] + f"{int(season[-2:]) + 1:02d}"
+            if season == "1920":
+                return "1919"
+            if season == "2021":
+                return "2020"
+            if season[:2] == "19" or season[:2] == "20":
+                return season
+            if int(season) <= current_year:
+                return "20" + season[:2]
+            return "19" + season[:2]
+
+        def process_two_digit_year(season: str) -> str:
+            """Process a 2-digit string like '94'."""
+            if self == SeasonCode.MULTI_YEAR:
+                if season == "99":
+                    return "9900"
+                return season + f"{int(season) + 1:02d}"
+            if int("20" + season) <= current_year:
+                return "20" + season
+            return "19" + season
+
+        def process_full_year_range(season: str) -> str:
+            """Process a range of 4-digit strings like '1994-1995'."""
+            if self == SeasonCode.MULTI_YEAR:
+                return season[2:4] + season[-2:]
+            return season[:4]
+
+        def process_partial_year_range(season: str) -> str:
+            """Process a range of 4-digit and 2-digit string like '1994-95'."""
+            if self == SeasonCode.MULTI_YEAR:
+                return season[2:4] + season[-2:]
+            return season[:4]
+
+        def process_short_year_range(season: str) -> str:
+            """Process a range of 2-digit strings like '94-95'."""
+            if self == SeasonCode.MULTI_YEAR:
+                return season[:2] + season[-2:]
+            if int("20" + season[:2]) <= current_year:
+                return "20" + season[:2]
+            return "19" + season[:2]
+
+        for pattern, action in patterns:
+            if pattern.match(season):
+                return action(season)
+
+        raise ValueError(f"Unrecognized season code: '{season}'")
 
 
 class BaseReader(ABC):
@@ -56,9 +228,9 @@ class BaseReader(ABC):
 
     def __init__(
         self,
-        leagues: Optional[Union[str, List[str]]] = None,
+        leagues: Optional[Union[str, list[str]]] = None,
         proxy: Optional[
-            Union[str, Dict[str, str], List[Dict[str, str]], Callable[[], Dict[str, str]]]
+            Union[str, dict[str, str], list[dict[str, str]], Callable[[], dict[str, str]]]
         ] = None,
         no_cache: bool = False,
         no_store: bool = False,
@@ -77,7 +249,7 @@ class BaseReader(ABC):
         elif callable(proxy):
             self.proxy = proxy
         else:
-            self.proxy = lambda: {}
+            self.proxy = dict
 
         self._selected_leagues = leagues  # type: ignore
         self.no_cache = no_cache
@@ -133,7 +305,8 @@ class BaseReader(ABC):
             logger.debug("Scraping %s", url)
             return self._download_and_save(url, filepath, var)
         logger.debug("Retrieving %s from cache", url)
-        assert filepath is not None
+        if filepath is None:
+            raise ValueError("No filepath provided for cached data.")
         return filepath.open(mode="rb")
 
     def _is_cached(
@@ -174,8 +347,8 @@ class BaseReader(ABC):
         cache_invalid = False
         # Check if cached file is too old
         if _max_age is not None and filepath is not None and filepath.exists():
-            last_modified = datetime.fromtimestamp(filepath.stat().st_mtime)
-            now = datetime.now()
+            last_modified = datetime.fromtimestamp(filepath.stat().st_mtime, tz=timezone.utc)
+            now = datetime.now(timezone.utc)
             if (now - last_modified) > _max_age:
                 cache_invalid = True
 
@@ -206,12 +379,12 @@ class BaseReader(ABC):
         """
 
     @classmethod
-    def available_leagues(cls) -> List[str]:
+    def available_leagues(cls) -> list[str]:
         """Return a list of league IDs available for this source."""
         return sorted(cls._all_leagues().keys())
 
     @classmethod
-    def _all_leagues(cls) -> Dict[str, str]:
+    def _all_leagues(cls) -> dict[str, str]:
         """Return a dict mapping all canonical league IDs to source league IDs."""
         if not hasattr(cls, "_all_leagues_dict"):
             cls._all_leagues_dict = {  # type: ignore
@@ -229,12 +402,12 @@ class BaseReader(ABC):
         return df
 
     @property
-    def _selected_leagues(self) -> Dict[str, str]:
+    def _selected_leagues(self) -> dict[str, str]:
         """Return a dict mapping selected canonical league IDs to source league IDs."""
         return self._leagues_dict
 
     @_selected_leagues.setter
-    def _selected_leagues(self, ids: Optional[Union[str, List[str]]] = None) -> None:
+    def _selected_leagues(self, ids: Optional[Union[str, list[str]]] = None) -> None:
         if ids is None:
             self._leagues_dict = self._all_leagues()
         else:
@@ -248,11 +421,15 @@ class BaseReader(ABC):
                     raise ValueError(
                         f"""
                         Invalid league '{i}'. Valid leagues are:
-                        { pprint.pformat(self.available_leagues()) }
+                        {pprint.pformat(self.available_leagues())}
                         """
                     )
                 tmp_league_dict[i] = self._all_leagues()[i]
             self._leagues_dict = tmp_league_dict
+
+    @property
+    def _season_code(self) -> SeasonCode:
+        return SeasonCode.from_leagues(self.leagues)
 
     def _is_complete(self, league: str, season: str) -> bool:
         """Check if a season is complete."""
@@ -265,22 +442,30 @@ class BaseReader(ABC):
             else:
                 raise ValueError(f"Invalid league '{league}'")
         if "season_end" not in league_dict:
-            season_ends = date(datetime.strptime(season[-2:], "%y").year, 7, 1)
-        else:
-            season_ends = date(
-                datetime.strptime(season[-2:], "%y").year,
-                datetime.strptime(league_dict["season_end"], "%b").month,
+            season_ends = datetime(
+                datetime.strptime(season[-2:], "%y").year,  # noqa: DTZ007
+                7,
                 1,
+                tzinfo=timezone.utc,
+            )
+        else:
+            season_ends = datetime(
+                datetime.strptime(season[-2:], "%y").year,  # noqa: DTZ007
+                datetime.strptime(  # noqa: DTZ007
+                    league_dict["season_end"], "%b"
+                ).month,
+                1,
+                tzinfo=timezone.utc,
             ) + relativedelta(months=1)
-        return date.today() >= season_ends
+        return datetime.now(tz=timezone.utc) >= season_ends
 
     @property
-    def leagues(self) -> List[str]:
+    def leagues(self) -> list[str]:
         """Return a list of selected leagues."""
         return list(self._leagues_dict.keys())
 
     @property
-    def seasons(self) -> List[str]:
+    def seasons(self) -> list[str]:
         """Return a list of selected seasons."""
         return self._season_ids
 
@@ -288,11 +473,11 @@ class BaseReader(ABC):
     def seasons(self, seasons: Optional[Union[str, int, Iterable[Union[str, int]]]]) -> None:
         if seasons is None:
             logger.info("No seasons provided. Will retrieve data for the last 5 seasons.")
-            year = datetime.today().year
-            seasons = [f"{y-1}-{y}" for y in range(year, year - 6, -1)]
-        if isinstance(seasons, str) or isinstance(seasons, int):
+            year = datetime.now(tz=timezone.utc).year
+            seasons = [f"{y - 1}-{y}" for y in range(year, year - 6, -1)]
+        if isinstance(seasons, (str, int)):
             seasons = [seasons]
-        self._season_ids = [season_code(s) for s in seasons]
+        self._season_ids = [self._season_code.parse(s) for s in seasons]
 
 
 class BaseRequestsReader(BaseReader):
@@ -300,9 +485,9 @@ class BaseRequestsReader(BaseReader):
 
     def __init__(
         self,
-        leagues: Optional[Union[str, List[str]]] = None,
+        leagues: Optional[Union[str, list[str]]] = None,
         proxy: Optional[
-            Union[str, Dict[str, str], List[Dict[str, str]], Callable[[], Dict[str, str]]]
+            Union[str, dict[str, str], list[dict[str, str]], Callable[[], dict[str, str]]]
         ] = None,
         no_cache: bool = False,
         no_store: bool = False,
@@ -320,15 +505,8 @@ class BaseRequestsReader(BaseReader):
         self._session = self._init_session()
 
     def _init_session(self) -> requests.Session:
-        session = requests.Session()
-        session.headers.update(
-            {
-                "User-Agent": (
-                    "Mozilla/5.0 (Windows NT 6.1; Win64; x64) "
-                    "AppleWebKit/537.36 (KHTML, like Gecko) "
-                    "Chrome/59.0.3071.86 Safari/537.36"
-                )
-            }
+        session = cloudscraper.create_scraper(
+            browser={"browser": "chrome", "platform": "linux", "mobile": False}
         )
         session.proxies.update(self.proxy())
         return session
@@ -349,7 +527,7 @@ class BaseRequestsReader(BaseReader):
                     if isinstance(var, str):
                         var = [var]
                     var_names = "|".join(var)
-                    template_understat = br"(%b)+[\s\t]*=[\s\t]*JSON\.parse\('(.*)'\)"
+                    template_understat = rb"(%b)+[\s\t]*=[\s\t]*JSON\.parse\('(.*)'\)"
                     pattern_understat = template_understat % bytes(var_names, encoding="utf-8")
                     results = re.findall(pattern_understat, response.content)
                     data = {
@@ -365,12 +543,14 @@ class BaseRequestsReader(BaseReader):
                 return io.BytesIO(payload)
             except Exception:
                 logger.exception(
-                    "Error while scraping %s. Retrying... (attempt %d of 5).", url, i + 1
+                    "Error while scraping %s. Retrying... (attempt %d of 5).",
+                    url,
+                    i + 1,
                 )
                 self._session = self._init_session()
                 continue
 
-        raise ConnectionError("Could not download %s." % url)
+        raise ConnectionError(f"Could not download {url}.")
 
 
 class BaseSeleniumReader(BaseReader):
@@ -378,9 +558,9 @@ class BaseSeleniumReader(BaseReader):
 
     def __init__(
         self,
-        leagues: Optional[Union[str, List[str]]] = None,
+        leagues: Optional[Union[str, list[str]]] = None,
         proxy: Optional[
-            Union[str, Dict[str, str], List[Dict[str, str]], Callable[[], Dict[str, str]]]
+            Union[str, dict[str, str], list[dict[str, str]], Callable[[], dict[str, str]]]
         ] = None,
         no_cache: bool = False,
         no_store: bool = False,
@@ -439,7 +619,7 @@ class BaseSeleniumReader(BaseReader):
             chrome_options.add_argument("--host-resolver-rules=" + resolver_rules)
         return uc.Chrome(options=chrome_options)
 
-    def _download_and_save(  # noqa: C901
+    def _download_and_save(
         self,
         url: str,
         filepath: Optional[Path] = None,
@@ -458,12 +638,17 @@ class BaseSeleniumReader(BaseReader):
                     response = self._driver.execute_script(
                         "return document.body.innerHTML;"
                     ).encode("utf-8")
+                    if response == b"":
+                        raise Exception("Empty response.")
                 else:
                     if not isinstance(var, str):
                         raise NotImplementedError("Only implemented for single variables.")
-                    response = json.dumps(self._driver.execute_script("return " + var)).encode(
-                        "utf-8"
-                    )
+                    try:
+                        response = json.dumps(self._driver.execute_script("return " + var)).encode(
+                            "utf-8"
+                        )
+                    except JavascriptException:
+                        response = json.dumps(None).encode("utf-8")
                 if not self.no_store and filepath is not None:
                     filepath.parent.mkdir(parents=True, exist_ok=True)
                     with filepath.open(mode="wb") as fh:
@@ -471,51 +656,16 @@ class BaseSeleniumReader(BaseReader):
                 return io.BytesIO(response)
             except Exception:
                 logger.exception(
-                    "Error while scraping %s. Retrying... (attempt %d of 5).", url, i + 1
+                    "Error while scraping %s. Retrying in %d seconds... (attempt %d of 5).",
+                    url,
+                    i * 10,
+                    i + 1,
                 )
+                time.sleep(i * 10)
                 self._driver = self._init_webdriver()
                 continue
 
-        raise ConnectionError("Could not download %s." % url)
-
-
-def season_code(season: Union[str, int]) -> str:  # noqa: C901
-    """Convert a string or int to a season code like '1718'."""
-    season = str(season)
-    pat1 = re.compile(r"^[0-9]{4}$")  # 1994 | 9495
-    pat2 = re.compile(r"^[0-9]{2}$")  # 94
-    pat3 = re.compile(r"^[0-9]{4}-[0-9]{4}$")  # 1994-1995
-    pat4 = re.compile(r"^[0-9]{4}/[0-9]{4}$")  # 1994/1995
-    pat5 = re.compile(r"^[0-9]{4}-[0-9]{2}$")  # 1994-95
-    pat6 = re.compile(r"^[0-9]{2}-[0-9]{2}$")  # 94-95
-
-    if re.match(pat1, season):
-        if int(season[2:]) == int(season[:2]) + 1:
-            if season == "2021":
-                msg = 'Season id "{}" is ambiguous: interpreting as "{}-{}"'.format(
-                    season, season[:2], season[-2:]
-                )
-                warnings.warn(msg, stacklevel=1)
-            return season  # 9495
-        elif season[2:] == "99":
-            return "".join([season[2:], "00"])  # 1999
-        else:
-            return "".join([season[-2:], f"{int(season[-2:]) + 1:02d}"])  # 1994
-    elif re.match(pat2, season):
-        if season == "99":
-            return "".join([season, "00"])  # 99
-        else:
-            return "".join([season, f"{int(season) + 1:02d}"])  # 94
-    elif re.match(pat3, season):
-        return "".join([season[2:4], season[-2:]])  # 1994-1995
-    elif re.match(pat4, season):
-        return "".join([season[2:4], season[-2:]])  # 1994/1995
-    elif re.match(pat5, season):
-        return "".join([season[2:4], season[-2:]])  # 1994-95
-    elif re.match(pat6, season):
-        return "".join([season[:2], season[-2:]])  # 94-95
-    else:
-        return season
+        raise ConnectionError(f"Could not download {url}.")
 
 
 def make_game_id(row: pd.Series) -> str:
@@ -534,7 +684,7 @@ def make_game_id(row: pd.Series) -> str:
     return game_id
 
 
-def standardize_colnames(df: pd.DataFrame, cols: Optional[List[str]] = None) -> pd.DataFrame:
+def standardize_colnames(df: pd.DataFrame, cols: Optional[list[str]] = None) -> pd.DataFrame:
     """Convert DataFrame column names to snake case."""
 
     def to_snake(name: str) -> str:
@@ -556,7 +706,7 @@ def standardize_colnames(df: pd.DataFrame, cols: Optional[List[str]] = None) -> 
     return df.rename(columns={c: to_snake(c) for c in cols})
 
 
-def get_proxy() -> Dict[str, str]:
+def get_proxy() -> dict[str, str]:
     """Return a public proxy."""
     # list of free proxy apis
     # protocols: http, https, socks4 and socks5
@@ -573,8 +723,7 @@ def get_proxy() -> Dict[str, str]:
         if not full_proxy_list:
             logger.info("There are currently no proxies available. Exiting...")
             return {}
-        else:
-            logger.info(f"Found {len(full_proxy_list)} proxy servers. Checking...\n")
+        logger.info(f"Found {len(full_proxy_list)} proxy servers. Checking...")
 
     # creating proxy dict
     final_proxy_list = []
@@ -605,5 +754,5 @@ def check_proxy(proxy: dict) -> bool:
         r0 = requests.get("https://ipinfo.io/json", proxies=proxy, timeout=15)
         return r0.status_code == 200
     except Exception as error:
-        logger.error(f"BAD PROXY: Reason: {str(error)}\n")
+        logger.error(f"BAD PROXY: Reason: {error!s}\n")
         return False
