@@ -1,17 +1,18 @@
 """A collection of tools to read and process soccer data from various sources."""
 
-__version__ = '1.6.0'
+__version__ = "1.8.1"
 
 __all__ = [
-    'ClubElo',
-    'ESPN',
-    'FBref',
-    'FiveThirtyEight',
-    'FotMob',
-    'MatchHistory',
-    'SoFIFA',
-    'Understat',
-    'WhoScored',
+    "ClubElo",
+    "ESPN",
+    "FBref",
+    "FiveThirtyEight",
+    "FotMob",
+    "MatchHistory",
+    "Sofascore",
+    "SoFIFA",
+    "Understat",
+    "WhoScored",
 ]
 
 from .clubelo import ClubElo
@@ -20,6 +21,7 @@ from .fbref import FBref
 from .fivethirtyeight import FiveThirtyEight
 from .fotmob import FotMob
 from .match_history import MatchHistory
+from .sofascore import Sofascore
 from .sofifa import SoFIFA
 from .understat import Understat
 from .whoscored import WhoScored
