@@ -19,7 +19,8 @@ def test_read_matchsheet(espn_seriea: ESPN) -> None:
 def test_read_matchsheet_bad_id(espn_seriea: ESPN) -> None:
     """It should raise a ValueError if the selected game is not in the specified season."""
     with pytest.raises(
-        ValueError, match="No games with the given IDs found for the selected seasons and leagues."
+        ValueError,
+        match="No games with the given IDs found for the selected seasons and leagues.",
     ):
         assert isinstance(espn_seriea.read_matchsheet(match_id=123), pd.DataFrame)
 
@@ -31,5 +32,8 @@ def test_read_lineups(espn_seriea: ESPN) -> None:
 
 def test_id_not_in_season(espn_seriea: ESPN) -> None:
     """It should raise a ValueError if the selected game is not in the specified season."""
-    with pytest.raises(ValueError):
+    with pytest.raises(
+        ValueError,
+        match="No games with the given IDs found for the selected seasons and leagues.",
+    ):
         assert isinstance(espn_seriea.read_lineup(match_id=123), pd.DataFrame)
