@@ -13,21 +13,21 @@ import soccerdata as foo
 # Scores per game equal for all common leagues over classes
 
 
-@pytest.mark.e2e
+@pytest.mark.e2e()
 def test_five38_vs_elo():
     """We should be able to retrieve the Elo history for all teams in these leagues."""
     league_sel = [
-        'ENG-Premier League',
-        'ESP-La Liga',
-        'FRA-Ligue 1',
-        'GER-Bundesliga',
-        'ITA-Serie A',
+        "ENG-Premier League",
+        "ESP-La Liga",
+        "FRA-Ligue 1",
+        "GER-Bundesliga",
+        "ITA-Serie A",
     ]
 
-    five38 = foo.FiveThirtyEight(leagues=league_sel, seasons='1819')
+    five38 = foo.FiveThirtyEight(leagues=league_sel, seasons="1819")
     five38_games = five38.read_games()
 
     elo = foo.ClubElo()
-    elo_hist = pd.concat([elo.read_team_history(team) for team in set(five38_games['home_team'])])
+    elo_hist = pd.concat([elo.read_team_history(team) for team in set(five38_games["home_team"])])
 
-    assert set(five38_games['home_team']) - set(elo_hist['team']) == set()
+    assert set(five38_games["home_team"]) - set(elo_hist["team"]) == set()
